requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
defusedxml==0.7.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
et-xmlfile==1.1.0
fastjsonschema==2.20.0
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.6.0
gast==0.5.4
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.15.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
opencv-python-headless==4.10.0.82
openpyxl==3.1.3
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
PyMatting==1.1.12
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
sympy==1.12.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tinycss2==1.3.0
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.14.1
